+++ assets/contao/js/core-uncompressed.js
@@ -2145,7 +2145,6 @@ var Backend =
 	 */
 	enableToggleSelect: function() {
 		var container = $('tl_select'),
-			checkboxes = [], start, thisIndex, startIndex, status, from, to,
 			shiftToggle = function(el) {
 				thisIndex = checkboxes.indexOf(el);
 				startIndex = checkboxes.indexOf(start);
@@ -2156,16 +2155,9 @@ var Backend =
 				for (from; from<=to; from++) {
 					checkboxes[from].checked = status;
 				}
-			};
-
-		if (container) {
-			checkboxes = container.getElements('input[type="checkbox"]');
-		}
-
-		// Row click
-		$$('.toggle_select').each(function(el) {
-			el.addEvent('click', function(e) {
-				var input = $(el).getElement('input[type="checkbox"],input[type="radio"]');
+			},
+			clickEvent = function(e) {
+				var input = this.getElement('input[type="checkbox"],input[type="radio"]');
 
 				if (!input) {
 					return;
@@ -2176,6 +2168,7 @@ var Backend =
 					if (!input.checked) {
 						input.checked = 'checked';
 					}
+
 					return;
 				}
 
@@ -2191,7 +2184,25 @@ var Backend =
 				}
 
 				start = input;
-			});
+			},
+			checkboxes = [], start, thisIndex, startIndex, status, from, to;
+
+		if (container) {
+			checkboxes = container.getElements('input[type="checkbox"]');
+		}
+
+		// Row click
+		$$('.toggle_select').each(function(el) {
+			var boundEvent = el.retrieve('boundEvent');
+
+			if (boundEvent) {
+				el.removeEvent('click', boundEvent);
+			}
+
+			boundEvent = clickEvent.bind(el);
+
+			el.addEvent('click', boundEvent);
+			el.store('boundEvent', boundEvent);
 		});
 
 		// Checkbox click